refactor: 홈 리듀서 책임을 파일별로 분리
