+++ Projects/Features/HomeFeature/Sources/Presentation/HomeFeatureReducer.swift
@@ -4,49 +4,6 @@ import DSKit
 import Foundation
 
 @Reducer
-struct HomeFilterReducer {
-    @ObservableState
-    struct State: Equatable, Sendable {
-        var regions: [RegionList] = []
-        var selectedRegion: RegionList?
-        var selectedDistrict: String?
-        var selectedOption: SortButton.SortOption = .newest
-    }
-
-    enum Action: Sendable {
-        case regionSelected(RegionList)
-        case districtSelected(String)
-        case sortOptionSelected(SortButton.SortOption)
-        case regionSelectionPrepared(HomeRegionSelection)
-    }
-
-    var body: some Reducer<State, Action> {
-        Reduce { state, action in
-            switch action {
-            case .regionSelected(let region):
-                state.selectedRegion = region
-                state.selectedDistrict = region.districtList.first
-                return .none
-
-            case .districtSelected(let district):
-                state.selectedDistrict = district
-                return .none
-
-            case .sortOptionSelected(let option):
-                state.selectedOption = option
-                return .none
-
-            case .regionSelectionPrepared(let selection):
-                state.regions = selection.regions
-                state.selectedRegion = selection.selectedRegion
-                state.selectedDistrict = selection.selectedDistrict
-                return .none
-            }
-        }
-    }
-}
-
-@Reducer
 struct HomeFeatureReducer {
     @ObservableState
     struct State: Equatable, Sendable {
@@ -275,67 +232,6 @@ struct HomePopupSections: Equatable, Sendable {
     var bestPopups: [Popup]
     var comingPopups: [Popup]
     var gridPopups: [Popup]
-}
-
-struct HomePopupClient: Sendable {
-    var getRegionList: @Sendable () async throws -> [RegionList]
-    var getPersonalRandomPopupList: @Sendable (_ userUuid: String) async throws -> [Popup]
-    var getPersonalUpcomingPopupList: @Sendable (_ userUuid: String) async throws -> [Popup]
-    var getPersonalFilteredPopupList: @Sendable (
-        _ userUuid: String,
-        _ region: String,
-        _ district: String,
-        _ homeSortStandard: String
-    ) async throws -> [Popup]
-    var addFavorite: @Sendable (_ userUuid: String, _ popupUuid: String) async throws -> Void
-    var removeFavorite: @Sendable (_ userUuid: String, _ popupUuid: String) async throws -> Void
-}
-
-extension HomePopupClient: DependencyKey {
-    static var liveValue: HomePopupClient {
-        let box = PopupUsecaseBox(DIContainer.shared.resolve(PopupUsecaseProtocol.self))
-
-        return HomePopupClient(
-            getRegionList: {
-                try await box.usecase.getRegionList()
-            },
-            getPersonalRandomPopupList: { userUuid in
-                try await box.usecase.getPersonalRandomPopupList(userUuid: userUuid)
-            },
-            getPersonalUpcomingPopupList: { userUuid in
-                try await box.usecase.getPersonalUpcomingPopupList(userUuid: userUuid)
-            },
-            getPersonalFilteredPopupList: { userUuid, region, district, homeSortStandard in
-                try await box.usecase.getPersonalFilteredPopupList(
-                    userUuid: userUuid,
-                    region: region,
-                    district: district,
-                    homeSortStandard: homeSortStandard
-                )
-            },
-            addFavorite: { userUuid, popupUuid in
-                try await box.usecase.addFavorite(userUuid: userUuid, popupUuid: popupUuid)
-            },
-            removeFavorite: { userUuid, popupUuid in
-                try await box.usecase.removeFavorite(userUuid: userUuid, popupUuid: popupUuid)
-            }
-        )
-    }
-}
-
-extension DependencyValues {
-    var homePopupClient: HomePopupClient {
-        get { self[HomePopupClient.self] }
-        set { self[HomePopupClient.self] = newValue }
-    }
-}
-
-private final class PopupUsecaseBox: @unchecked Sendable {
-    let usecase: PopupUsecaseProtocol
-
-    init(_ usecase: PopupUsecaseProtocol) {
-        self.usecase = usecase
-    }
 }
 
 private extension [RegionList] {
